feat: display topic on proposal card and refactor meta
