+++ frontend/svelte/src/lib/components/proposal-detail/ProposalDetailCard/ProposalDetailCard.svelte
@@ -6,7 +6,7 @@
   import type { ProposalColor } from "../../../../lib/constants/proposals.constants";
   import { PROPOSAL_COLOR } from "../../../../lib/constants/proposals.constants";
   import { i18n } from "../../../../lib/stores/i18n";
-  import ProposalMeta from "./ProposalMeta.svelte";
+  import ProposalMeta from "../../proposals/ProposalMeta.svelte";
   import ProposalActions from "./ProposalActions.svelte";
   import ProposalSummaryCardBlock from "./ProposalSummaryCardBlock.svelte";
 
@@ -28,7 +28,11 @@
     ><h2 class="status">{$i18n.status[ProposalStatus[status]]}</h2></Badge
   >
   <ProposalSummaryCardBlock {proposal} />
-  <ProposalMeta {proposalInfo} />
+
+  <div class="detail">
+    <ProposalMeta {proposalInfo} />
+  </div>
+
   <ProposalActions {proposal} />
 </Card>
 
@@ -53,5 +57,9 @@
     min-width: fit-content;
     font-size: var(--font-size-h3);
     color: inherit;
+  }
+
+  .detail {
+    margin: var(--padding-3x) 0;
   }
 </style>
