+++ frontend/svelte/src/lib/types/i18n.d.ts
@@ -303,7 +303,6 @@ interface I18nProposal_detail__ineligible {
 
 interface I18nNeuron_detail {
   title: string;
-  proposer: string;
   voting_history: string;
   vote: string;
   join_community_fund: string;
