+++ frontend/svelte/src/tests/lib/components/proposals/Proposer.spec.ts
@@ -36,7 +36,7 @@ describe("Proposer", () => {
     });
 
     expect(container.querySelector("button small")?.textContent).toEqual(
-      `${en.neuron_detail.proposer}: ${mockProposalInfo.proposer}`
+      `${en.proposal_detail.proposer_prefix} ${mockProposalInfo.proposer}`
     );
   });
